# Gateway/Adapters/Http/Contract/employeebffopenapi.yaml
openapi: 3.0.0
info:
  title: Employee BFF
  version: 1.0.0
  description: BFF для работы с приложениями сотрудников
...
paths:
...
  /v1.0/rents:
    get:
      tags:
        - rent
      summary: Get All Current Rents
      operationId: getAllCurrentRents
      description: Получает список всех текущих аренд.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            format: int32
          required: true
          description: Номер страницы для пагинации.
        - in: query
          name: pageSize
          schema:
            type: integer
            format: int32
          required: true
          description: Количество элементов на странице для пагинации.
      responses:
        '200':
          description: Успешный ответ со списком текущих аренд.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllCurrentRentsResponse'
        '400':
          description: Некорректные параметры пагинации.
  /v1.0/rents/{rentId}:
    get:
      tags:
        - rent
      summary: Get Rent by ID
      operationId: getRentById
      description: Получает полную информацию об аренде по её уникальному идентификатору.
      parameters:
        - in: path
          name: rentId
          schema:
            type: string
            format: uuid
          required: true
          description: Уникальный идентификатор аренды.
      responses:
        '200':
          description: Успешный ответ с информацией об аренде.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetRentByIdResponse'
        '404':
          description: Аренда с указанным ID не найдена.
  /v1.0/rents/{rentId}/current-amount:
    get:
      tags:
        - rent
      summary: Get Current Amount Rent
      operationId: getCurrentAmountRent
      description: Получает текущую стоимость аренды по её уникальному идентификатору.
      parameters:
        - in: path
          name: rentId
          schema:
            type: string
            format: uuid
          required: true
          description: Уникальный идентификатор аренды.
      responses:
        '200':
          description: Успешный ответ с текущей стоимостью аренды.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCurrentAmountRentResponse'
        '404':
          description: Аренда с указанным ID не найдена.
  /v1.0/driving-licenses/{licenseId}:
    get:
      tags:
        - driving-license
      summary: Get Driving License by ID
      operationId: getLicenseById
      description: Получает информацию о водительском удостоверении по его уникальному идентификатору.
      parameters:
        - in: path
          name: licenseId
          schema:
            type: string
            format: uuid
          required: true
          description: Уникальный идентификатор водительского удостоверения.
      responses:
        '200':
          description: Успешный ответ с информацией о водительском удостоверении.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLicenseByIdResponse'
        '404':
          description: Водительское удостоверение с указанным ID не найдено.
  /v1.0/driving-licenses:
    get:
      tags:
        - driving-license
      summary: Get All Driving Licenses
      operationId: getAllLicenses
      description: Получает список всех водительских удостоверений с возможностью фильтрации по статусу.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            format: int32
          required: false
          description: Номер страницы для пагинации.
        - in: query
          name: pageSize
          schema:
            type: integer
            format: int32
          required: false
          description: Количество элементов на странице для пагинации.
        - in: query
          name: filteringStatus
          schema:
            $ref: '#/components/schemas/DrivingLicenseStatus'
          required: true
          description: Статус для фильтрации водительских удостоверений.
      responses:
        '200':
          description: Успешный ответ со списком водительских удостоверений.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAllLicensesResponse'
        '400':
          description: Некорректные параметры запроса.
  /v1.0/driving-licenses/{licenseId}/approve:
    post:
      tags:
        - driving-license
      summary: Approve Driving License
      operationId: approveLicense
      description: Одобряет водительское удостоверение.
      parameters:
        - in: path
          name: licenseId
          schema:
            type: string
            format: uuid
          required: true
          description: Уникальный идентификатор водительского удостоверения.
      responses:
        '200':
          description: Водительское удостоверение успешно одобрено.
        '404':
          description: Водительское удостоверение с указанным ID не найдено.
        '400':
          description: Невозможно одобрить водительское удостоверение в текущем статусе.
  /v1.0/driving-licenses/{licenseId}/reject:
    post:
      tags:
        - driving-license
      summary: Reject Driving License
      operationId: rejectLicense
      description: Отклоняет водительское удостоверение.
      parameters:
        - in: path
          name: licenseId
          schema:
            type: string
            format: uuid
          required: true
          description: Уникальный идентификатор водительского удостоверения.
      responses:
        '200':
          description: Водительское удостоверение успешно отклонено.
        '404':
          description: Водительское удостоверение с указанным ID не найдено.
        '400':
          description: Невозможно отклонить водительское удостоверение в текущем статусе.
components:
  schemas:
...
    GetAllCurrentRentsResponse:
      type: object
      required:
        - rents
      properties:
        rents:
          type: array
          items:
            $ref: '#/components/schemas/Rent'
          description: Список текущих аренд.
    Rent:
      type: object
      required:
        - rentId
        - vehicleId
        - customerId
        - start
      properties:
        rentId:
          type: string
          format: uuid
          description: Уникальный идентификатор аренды.
        vehicleId:
          type: string
          format: uuid
          description: Уникальный идентификатор автомобиля.
        customerId:
          type: string
          format: uuid
          description: Уникальный идентификатор работника.
        start:
          type: string
          format: date-time
          description: Дата и время начала аренды.
    RentStatus:
      type: string
      enum:
        - inProgress
        - completed
      description: Статус аренды.
    GetRentByIdResponse:
      type: object
      required:
        - rentId
        - bookingId
        - vehicleId
        - customerId
        - start
        - rentStatus
      properties:
        rentId:
          type: string
          format: uuid
          description: Уникальный идентификатор аренды.
        vehicleId:
          type: string
          format: uuid
          description: Уникальный идентификатор автомобиля.
        bookingId:
          type: string
          format: uuid
          description: Уникальный идентификатор бронирования.
        customerId:
          type: string
          format: uuid
          description: Уникальный идентификатор работника.
        start:
          type: string
          format: date-time
          description: Дата и время начала аренды.
        end:
          type: string
          format: date-time
          description: Дата и время окончания аренды.
        actualAmount:
          type: number
          format: double
          description: Общая стоимость аренды.
        rentStatus:
          $ref: '#/components/schemas/RentStatus'
    GetCurrentAmountRentResponse:
      type: object
      required:
        - currentAmount
      properties:
        currentAmount:
          type: number
          format: double
          description: Текущая стоимость аренды.
    GetLicenseByIdResponse:
      type: object
      required:
        - id
        - accId
        - status
        - categories
        - number
        - name
        - cityOfBirth
        - dateOfBirth
        - codeOfIssue
        - dateOfIssue
        - dateOfExpiry
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор водительского удостоверения
        customerId:
          type: string
          format: uuid
          description: Идентификатор аккаунта владельца
        status:
          $ref: '#/components/schemas/DrivingLicenseStatus'
        categories:
          type: array
          items:
            type: string
            maxLength: 1
            example: 'B'
          description: Категории водительского удостоверения
        number:
          type: string
          description: Номер водительского удостоверения
        name:
          type: string
          description: ФИО владельца
        cityOfBirth:
          type: string
          description: Город рождения
        dateOfBirth:
          type: string
          format: date
          description: Дата рождения
        codeOfIssue:
          type: string
          description: Код подразделения выдачи
        dateOfIssue:
          type: string
          format: date
          description: Дата выдачи
        dateOfExpiry:
          type: string
          format: date
          description: Дата окончания действия
        frontPhotoS3Url:
          type: string
          description: URL фотографии лицевой стороны в S3
          nullable: true
        backPhotoS3Url:
          type: string
          description: URL фотографии обратной стороны в S3
          nullable: true
    GetAllLicensesResponse:
      type: object
      required:
        - drivingLicenses
        - totalCount
      properties:
        drivingLicenses:
          type: array
          items:
            $ref: '#/components/schemas/DrivingLicenseShortView'
          description: Список всех водительских удостоверений.
    DrivingLicenseShortView:
      type: object
      required:
        - licenseId
        - customerId
        - status
      properties:
        licenseId:
          type: string
          format: uuid
          description: Уникальный идентификатор водительского удостоверения.
        customerId:
          type: string
          format: uuid
          description: Уникальный идентификатор работника, которому принадлежит удостоверение.
        status:
          $ref: '#/components/schemas/DrivingLicenseStatus'
    DrivingLicenseStatus:
      type: string
      enum:
        - pendingPhotosAdding
        - pendingProcessing
        - approved
        - rejected
        - expired
      description: Статус водительского удостоверения.
...