# Gateway/Adapters/Http/Contract/src/OpenApiContractV1/Controllers/DrivingLicenseApi.cs
/*
 * Employee BFF
 *
 * BFF для работы с приложениями сотрудников
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * Generated by: https://openapi-generator.tech
 */

using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Mvc;
using OpenApiContractV1.Attributes;
using OpenApiContractV1.Models;
using Swashbuckle.AspNetCore.Annotations;

namespace OpenApiContractV1.Controllers
{
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class DrivingLicenseApiController : ControllerBase
    {
        /// <summary>
        /// Approve Driving License
        /// </summary>
        /// <remarks>Одобряет водительское удостоверение.</remarks>
        /// <param name="licenseId">Уникальный идентификатор водительского удостоверения.</param>
        /// <response code="200">Водительское удостоверение успешно одобрено.</response>
        /// <response code="404">Водительское удостоверение с указанным ID не найдено.</response>
        /// <response code="400">Невозможно одобрить водительское удостоверение в текущем статусе.</response>
        [HttpPost]
        [Route("/v1.0/driving-licenses/{licenseId}/approve")]
        [ValidateModelState]
        [SwaggerOperation("ApproveLicense")]
        public abstract Task<IActionResult> ApproveLicense([FromRoute(Name = "licenseId")] [Required] Guid licenseId);

        /// <summary>
        /// Get All Driving Licenses
        /// </summary>
        /// <remarks>Получает список всех водительских удостоверений с возможностью фильтрации по статусу.</remarks>
        /// <param name="filteringStatus">Статус для фильтрации водительских удостоверений.</param>
        /// <param name="page">Номер страницы для пагинации.</param>
        /// <param name="pageSize">Количество элементов на странице для пагинации.</param>
        /// <response code="200">Успешный ответ со списком водительских удостоверений.</response>
        /// <response code="400">Некорректные параметры запроса.</response>
        [HttpGet]
        [Route("/v1.0/driving-licenses")]
        [ValidateModelState]
        [SwaggerOperation("GetAllLicenses")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetAllLicensesResponse),
            description: "Успешный ответ со списком водительских удостоверений.")]
        public abstract Task<IActionResult> GetAllLicenses(
            [FromQuery(Name = "filteringStatus")] [Required()] DrivingLicenseStatus filteringStatus,
            [FromQuery(Name = "page")] int? page,
            [FromQuery(Name = "pageSize")] int? pageSize);

        /// <summary>
        /// Get Driving License by ID
        /// </summary>
        /// <remarks>Получает информацию о водительском удостоверении по его уникальному идентификатору.</remarks>
        /// <param name="licenseId">Уникальный идентификатор водительского удостоверения.</param>
        /// <response code="200">Успешный ответ с информацией о водительском удостоверении.</response>
        /// <response code="404">Водительское удостоверение с указанным ID не найдено.</response>
        [HttpGet]
        [Route("/v1.0/driving-licenses/{licenseId}")]
        [ValidateModelState]
        [SwaggerOperation("GetLicenseById")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetLicenseByIdResponse),
            description: "Успешный ответ с информацией о водительском удостоверении.")]
        public abstract Task<IActionResult> GetLicenseById([FromRoute(Name = "licenseId")] [Required] Guid licenseId);

        /// <summary>
        /// Reject Driving License
        /// </summary>
        /// <remarks>Отклоняет водительское удостоверение.</remarks>
        /// <param name="licenseId">Уникальный идентификатор водительского удостоверения.</param>
        /// <response code="200">Водительское удостоверение успешно отклонено.</response>
        /// <response code="404">Водительское удостоверение с указанным ID не найдено.</response>
        /// <response code="400">Невозможно отклонить водительское удостоверение в текущем статусе.</response>
        [HttpPost]
        [Route("/v1.0/driving-licenses/{licenseId}/reject")]
        [ValidateModelState]
        [SwaggerOperation("RejectLicense")]
        public abstract Task<IActionResult> RejectLicense([FromRoute(Name = "licenseId")] [Required] Guid licenseId);
    }
}