# Gateway/Adapters/Http/Contract/src/OpenApiContractV1/Controllers/RentApi.cs
/*
 * Employee BFF
 *
 * BFF для работы с приложениями сотрудников
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * Generated by: https://openapi-generator.tech
 */

using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Mvc;
using OpenApiContractV1.Attributes;
using OpenApiContractV1.Models;
using Swashbuckle.AspNetCore.Annotations;

namespace OpenApiContractV1.Controllers
{
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public abstract class RentApiController : ControllerBase
    {
        /// <summary>
        /// Get All Current Rents
        /// </summary>
        /// <remarks>Получает список всех текущих аренд.</remarks>
        /// <param name="page">Номер страницы для пагинации.</param>
        /// <param name="pageSize">Количество элементов на странице для пагинации.</param>
        /// <response code="200">Успешный ответ со списком текущих аренд.</response>
        /// <response code="400">Некорректные параметры пагинации.</response>
        [HttpGet]
        [Route("/v1.0/rents")]
        [ValidateModelState]
        [SwaggerOperation("GetAllCurrentRents")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetAllCurrentRentsResponse),
            description: "Успешный ответ со списком текущих аренд.")]
        public abstract Task<IActionResult> GetAllCurrentRents(
            [FromQuery(Name = "page")] [Required()] int page,
            [FromQuery(Name = "pageSize")] [Required()] int pageSize);

        /// <summary>
        /// Get Current Amount Rent
        /// </summary>
        /// <remarks>Получает текущую стоимость аренды по её уникальному идентификатору.</remarks>
        /// <param name="rentId">Уникальный идентификатор аренды.</param>
        /// <response code="200">Успешный ответ с текущей стоимостью аренды.</response>
        /// <response code="404">Аренда с указанным ID не найдена.</response>
        [HttpGet]
        [Route("/v1.0/rents/{rentId}/current-amount")]
        [ValidateModelState]
        [SwaggerOperation("GetCurrentAmountRent")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetCurrentAmountRentResponse),
            description: "Успешный ответ с текущей стоимостью аренды.")]
        public abstract Task<IActionResult> GetCurrentAmountRent([FromRoute(Name = "rentId")] [Required] Guid rentId);

        /// <summary>
        /// Get Rent by ID
        /// </summary>
        /// <remarks>Получает полную информацию об аренде по её уникальному идентификатору.</remarks>
        /// <param name="rentId">Уникальный идентификатор аренды.</param>
        /// <response code="200">Успешный ответ с информацией об аренде.</response>
        /// <response code="404">Аренда с указанным ID не найдена.</response>
        [HttpGet]
        [Route("/v1.0/rents/{rentId}")]
        [ValidateModelState]
        [SwaggerOperation("GetRentById")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetRentByIdResponse),
            description: "Успешный ответ с информацией об аренде.")]
        public abstract Task<IActionResult> GetRentById([FromRoute(Name = "rentId")] [Required] Guid rentId);
    }
}