# Gateway/Adapters/Http/Contract/src/OpenApiContractV1/Models/DrivingLicenseShortView.cs
/*
 * Employee BFF
 *
 * BFF для работы с приложениями сотрудников
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * Generated by: https://openapi-generator.tech
 */

using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text;
using Newtonsoft.Json;

namespace OpenApiContractV1.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class DrivingLicenseShortView : IEquatable<DrivingLicenseShortView>
    {
        /// <summary>
        /// Уникальный идентификатор водительского удостоверения.
        /// </summary>
        /// <value>Уникальный идентификатор водительского удостоверения.</value>
        [Required]
        [DataMember(Name = "licenseId", EmitDefaultValue = true)]
        public Guid LicenseId { get; set; }

        /// <summary>
        /// Уникальный идентификатор работника, которому принадлежит удостоверение.
        /// </summary>
        /// <value>Уникальный идентификатор работника, которому принадлежит удостоверение.</value>
        [Required]
        [DataMember(Name = "customerId", EmitDefaultValue = true)]
        public Guid CustomerId { get; set; }

        /// <summary>
        /// Gets or Sets Status
        /// </summary>
        [Required]
        [DataMember(Name = "status", EmitDefaultValue = true)]
        public DrivingLicenseStatus Status { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class DrivingLicenseShortView {\n");
            sb.Append("  LicenseId: ").Append(LicenseId).Append("\n");
            sb.Append("  CustomerId: ").Append(CustomerId).Append("\n");
            sb.Append("  Status: ").Append(Status).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((DrivingLicenseShortView)obj);
        }

        /// <summary>
        /// Returns true if DrivingLicenseShortView instances are equal
        /// </summary>
        /// <param name="other">Instance of DrivingLicenseShortView to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(DrivingLicenseShortView other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return
                (
                    LicenseId == other.LicenseId ||
                    LicenseId.Equals(other.LicenseId)
                ) &&
                (
                    CustomerId == other.CustomerId ||
                    CustomerId.Equals(other.CustomerId)
                ) &&
                (
                    Status == other.Status ||
                    Status.Equals(other.Status)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)

                hashCode = hashCode * 59 + LicenseId.GetHashCode();

                hashCode = hashCode * 59 + CustomerId.GetHashCode();

                hashCode = hashCode * 59 + Status.GetHashCode();
                return hashCode;
            }
        }

        #region Operators

#pragma warning disable 1591

        public static bool operator ==(DrivingLicenseShortView left, DrivingLicenseShortView right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(DrivingLicenseShortView left, DrivingLicenseShortView right)
        {
            return !Equals(left, right);
        }

#pragma warning restore 1591

        #endregion Operators
    }
}