# Gateway/Adapters/Http/Contract/src/OpenApiContractV1/Models/GetAllLicensesResponse.cs
/*
 * Employee BFF
 *
 * BFF для работы с приложениями сотрудников
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * Generated by: https://openapi-generator.tech
 */

using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text;
using Newtonsoft.Json;

namespace OpenApiContractV1.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class GetAllLicensesResponse : IEquatable<GetAllLicensesResponse>
    {
        /// <summary>
        /// Список всех водительских удостоверений.
        /// </summary>
        /// <value>Список всех водительских удостоверений.</value>
        [Required]
        [DataMember(Name = "drivingLicenses", EmitDefaultValue = false)]
        public List<DrivingLicenseShortView> DrivingLicenses { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class GetAllLicensesResponse {\n");
            sb.Append("  DrivingLicenses: ").Append(DrivingLicenses).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((GetAllLicensesResponse)obj);
        }

        /// <summary>
        /// Returns true if GetAllLicensesResponse instances are equal
        /// </summary>
        /// <param name="other">Instance of GetAllLicensesResponse to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(GetAllLicensesResponse other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return
            (
                DrivingLicenses == other.DrivingLicenses ||
                DrivingLicenses != null &&
                other.DrivingLicenses != null &&
                DrivingLicenses.SequenceEqual(other.DrivingLicenses)
            );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                if (DrivingLicenses != null)
                    hashCode = hashCode * 59 + DrivingLicenses.GetHashCode();
                return hashCode;
            }
        }

        #region Operators

#pragma warning disable 1591

        public static bool operator ==(GetAllLicensesResponse left, GetAllLicensesResponse right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(GetAllLicensesResponse left, GetAllLicensesResponse right)
        {
            return !Equals(left, right);
        }

#pragma warning restore 1591

        #endregion Operators
    }
}