# Gateway/Adapters/Http/Contract/src/OpenApiContractV1/Models/Rent.cs
/*
 * Employee BFF
 *
 * BFF для работы с приложениями сотрудников
 *
 * The version of the OpenAPI document: 1.0.0
 *
 * Generated by: https://openapi-generator.tech
 */

using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using System.Text;
using Newtonsoft.Json;

namespace OpenApiContractV1.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class Rent : IEquatable<Rent>
    {
        /// <summary>
        /// Уникальный идентификатор аренды.
        /// </summary>
        /// <value>Уникальный идентификатор аренды.</value>
        [Required]
        [DataMember(Name = "rentId", EmitDefaultValue = true)]
        public Guid RentId { get; set; }

        /// <summary>
        /// Уникальный идентификатор автомобиля.
        /// </summary>
        /// <value>Уникальный идентификатор автомобиля.</value>
        [Required]
        [DataMember(Name = "vehicleId", EmitDefaultValue = true)]
        public Guid VehicleId { get; set; }

        /// <summary>
        /// Уникальный идентификатор работника.
        /// </summary>
        /// <value>Уникальный идентификатор работника.</value>
        [Required]
        [DataMember(Name = "customerId", EmitDefaultValue = true)]
        public Guid CustomerId { get; set; }

        /// <summary>
        /// Дата и время начала аренды.
        /// </summary>
        /// <value>Дата и время начала аренды.</value>
        [Required]
        [DataMember(Name = "start", EmitDefaultValue = true)]
        public DateTime Start { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class Rent {\n");
            sb.Append("  RentId: ").Append(RentId).Append("\n");
            sb.Append("  VehicleId: ").Append(VehicleId).Append("\n");
            sb.Append("  CustomerId: ").Append(CustomerId).Append("\n");
            sb.Append("  Start: ").Append(Start).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((Rent)obj);
        }

        /// <summary>
        /// Returns true if Rent instances are equal
        /// </summary>
        /// <param name="other">Instance of Rent to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(Rent other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return
                (
                    RentId == other.RentId ||
                    RentId.Equals(other.RentId)
                ) &&
                (
                    VehicleId == other.VehicleId ||
                    VehicleId.Equals(other.VehicleId)
                ) &&
                (
                    CustomerId == other.CustomerId ||
                    CustomerId.Equals(other.CustomerId)
                ) &&
                (
                    Start == other.Start ||
                    Start.Equals(other.Start)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)

                hashCode = hashCode * 59 + RentId.GetHashCode();

                hashCode = hashCode * 59 + VehicleId.GetHashCode();

                hashCode = hashCode * 59 + CustomerId.GetHashCode();

                hashCode = hashCode * 59 + Start.GetHashCode();
                return hashCode;
            }
        }

        #region Operators

#pragma warning disable 1591

        public static bool operator ==(Rent left, Rent right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(Rent left, Rent right)
        {
            return !Equals(left, right);
        }

#pragma warning restore 1591

        #endregion Operators
    }
}